% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_blocks.R
\name{bkd}
\alias{bkd}
\title{Get block diagonal matrix}
\usage{
bkd(df)
}
\arguments{
\item{df}{A block matrix with named rows and columns. Names of countries
and sectors are automatically identified.}
}
\value{
Block diagonal version of the original matrix.
}
\description{
Produces a block diagonal matrix version of block matrix, i.e., a matrix
in which the diagonal blocks are non-zero and the off-diagonal blocks
are zero
}
\examples{
wio <- make_wio("wiodtest", quiet = TRUE)
# Block diagonal version of Y (coincides with Yd)
bkd(wio$Y)
}
