% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_blocks.R
\name{bkoffd}
\alias{bkoffd}
\title{Get block off-diagonal matrix}
\usage{
bkoffd(df)
}
\arguments{
\item{df}{A block matrix with named rows and columns. Names of countries
and sectors are automatically identified.}
}
\value{
Block off-diagonal version of the original matrix.
}
\description{
Produces a block off-diagonal matrix version of block matrix, i.e., a matrix
in which the diagonal blocks are zero and the off-diagonal blocks
are non-zero.
}
\examples{
wio <- make_wio("wiodtest", quiet = TRUE)
# Block off-diagonal version of Y (coincides with Ym)
bkoffd(wio$Y)
}
