% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_mult.R, R/matrix_sum.R
\name{diagcs}
\alias{diagcs}
\title{Diagonalize the sums of columns of a matrix}
\usage{
diagcs(df)

diagcs(df)
}
\arguments{
\item{df}{A matrix with named rows and columns.}
}
\value{
A diagonal matrix with the sums of columns in the diagonal.

A diagonal matrix with the sums of columns in the diagonal.
}
\description{
Makes a diagonal matrix with the sums of columns of a matrix.

Diagonalizes the sums of each column of a matrix.
}
\examples{
wio <- make_wio("wiodtest")
diagcs(wio$W \%*\% wio$Bd)
wio <- make_wio("wiodtest")
diagcs(wio$W \%*\% wio$Bd)
}
