% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_sum.R
\name{diagrs}
\alias{diagrs}
\title{Diagonalize the sums of each row of a matrix}
\usage{
diagrs(df)
}
\arguments{
\item{df}{A matrix with named rows and columns.}
}
\value{
A diagonal matrix with the sums of rows in the diagonal.
}
\description{
Diagonalizes the sums of each row of a matrix.
}
\examples{
wio <- make_wio("wiodtest")
diagrs(wio$Y)
}
