% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iciotest_data}
\alias{iciotest_data}
\title{ICIO-type input-output table example data}
\format{
A matrix of 30 by 42, composed of two sub-matrices:
\itemize{
\item Sub-matrix \code{Z}, intermediate inputs: 30 by 30 (10 countries with 3
sectors each). 4 of those 10 are the extensions of \code{CHN} and \code{MEX}).
\item Sub-matrix \code{Yfd}, final demand: 30 by 12 (10 countries with 3 sectors
each in rows, 6 countries by 2 demand components each in columns).
}
}
\source{
Data were randomly generated with an uniform distribution.
}
\usage{
iciotest_data
}
\description{
An example of an ICIO-type input-output table, with rows for \code{MEX} and \code{CHN}
disaggregated into \code{MX1} and \code{MX2} and \code{CN1} and \code{CN2}, respectively.
}
\keyword{datasets}
