% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_geosec.R
\name{info_geo}
\alias{info_geo}
\title{Show available countries and country groups in a specific Input-Output table}
\usage{
info_geo(wiotype = "icio2025", lang = "en")
}
\arguments{
\item{wiotype}{Character string specifying the world input-output database.}

\item{lang}{Character string for the language of the descriptive text:
\code{"eng"} for English (default) and \code{"es"} for Spanish.}
}
\value{
Prints country codes and descriptive text in the console.
}
\description{
Show available countries and country groups in a specific Input-Output table
}
\examples{
info_geo("icio2025")
}
