% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meld.R
\name{meld}
\alias{meld}
\title{Meld ICIO-type matrix (consolidating China and Mexico sub-components)}
\usage{
meld(df, meld_rows = TRUE, meld_cols = TRUE)
}
\arguments{
\item{df}{A block matrix.}

\item{meld_rows}{Boolean, true to meld rows.}

\item{meld_cols}{Boolean, true to meld cols.}
}
\value{
Melded version of \code{ICIO} matrix.
}
\description{
Meld \code{ICIO} matrix with extended countries.
Melds countries \code{CHN} and \code{MEX} from their extended versions
e.g., \code{CN1} and \code{CN2} are melded into \code{CHN}.
}
