% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.exvadec}
\alias{print.exvadec}
\title{Print method for \code{exvadec} class}
\usage{
\method{print}{exvadec}(x, ...)
}
\arguments{
\item{x}{An object of class \code{exvadec}}

\item{...}{Additional arguments}
}
\value{
Printout to console
}
\description{
Print method for \code{exvadec} class
}
