% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.exvadir}
\alias{print.exvadir}
\title{Print method for \code{exvadir} class}
\usage{
\method{print}{exvadir}(x, ...)
}
\arguments{
\item{x}{An object of class \code{exvadir}}

\item{...}{Additional arguments}
}
\value{
Printout to console
}
\description{
Print method for \code{exvadir} class
}
