% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.wio}
\alias{print.wio}
\title{Print method for \code{wio} class}
\usage{
\method{print}{wio}(x, ...)
}
\arguments{
\item{x}{An object of class \code{wio}}

\item{...}{Additional arguments}
}
\value{
Printout to console
}
\description{
Print method for \code{wio} class
}
