% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.exvadec}
\alias{summary.exvadec}
\title{Summary method for \code{exvadec} class}
\usage{
\method{summary}{exvadec}(object, ...)
}
\arguments{
\item{object}{An object of class \code{exvadec}.}

\item{...}{Additional arguments.}
}
\value{
Printout to console
}
\description{
Summary method for \code{exvadec} class
}
