% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.exvadir}
\alias{summary.exvadir}
\title{Summary method for \code{exvadir} class}
\usage{
\method{summary}{exvadir}(object, ...)
}
\arguments{
\item{object}{An object of class \code{exvadir}.}

\item{...}{Additional arguments.}
}
\value{
Printout to console
}
\description{
Summary method for \code{exvadir} class
}
