% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.wio}
\alias{summary.wio}
\title{Summary method for \code{wio} class}
\usage{
\method{summary}{wio}(object, ...)
}
\arguments{
\item{object}{An object of class \code{wio}}

\item{...}{Additional arguments}
}
\value{
Printout to console
}
\description{
Summary method for \code{wio} class
}
