% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiodtest_data}
\alias{wiodtest_data}
\title{WIOD-type input-output table example data}
\format{
A matrix of 18 by 30, composed of two sub-matrices:
\itemize{
\item Sub-matrix \code{Z}, intermediate inputs: 18 by 18 (10 countries with 3
sectors each).
\item Sub-matrix \code{Yfd}, final demand: 18 by 12 (6 countries with 3 sectors
each in rows, 6 countries by 2 demand components each in columns).
}
}
\source{
Data were randomly generated with an uniform distribution.
}
\usage{
wiodtest_data
}
\description{
An example of a WIOD-type input-output table.
}
\keyword{datasets}
