% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_geosec.R
\name{info_sec}
\alias{info_sec}
\title{Show available sectors and sector groups included in a specific Input-Output table}
\usage{
info_sec(wiotype = "icio2021", lang = "en")
}
\arguments{
\item{wiotype}{Character string specifying the world input-output database}

\item{lang}{Character string for the language of the descriptive text:
\code{eng} for English (default) and \code{es} for Spanish.}
}
\value{
Prints ids, sector codes and descriptive text
}
\description{
Show available sectors and sector groups included in a specific Input-Output table
}
\examples{
info_sec("icio2021")
}
